+++ src/constants/routes.ts
@@ -13,5 +13,3 @@ export const WRITE_ARTICLE_PAGE = '/write_article';
 export const CONTACT = '/contact';
 export const ONBOARDING = '/onboarding';
 export const RESOURCES = '/resources';
-
-
